pyproject.toml
[project]
name = "file-catalog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "beautifulsoup4>=4.14.3",
    "faster-whisper>=1.2.1",
    "greenlet>=3.3.2",
    "huggingface-hub>=1.5.0",
    "langdetect>=1.0.9",
    "llama-cpp-python>=0.3.16",
    "odfpy>=1.4.1",
    "opencv-python-headless>=4.13.0.92",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.1",
    "pre-commit>=4.5.1",
    "psutil>=7.2.2",
    "pytesseract>=0.3.13",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-docx>=1.2.0",
    "python-magic>=0.4.27",
    "rich>=14.3.3",
    "ruff>=0.15.2",
    "sqlmodel>=0.0.37",
    "streamlit>=1.54.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
